spring24/Team 8/graphic-project-master/requirements.txt
PyOpenGL
pygame
numpy